feat: refreshToken auto fetching

+++ frontend/src/routes/Index.tsx
@@ -3,21 +3,21 @@ import { useEffect } from 'react';
 import { getTodoList } from '../hooks/useGetTodoList';
 import { receivedTodos } from '../../types';
 import apiStorage from '../utils/apiStorage';
+import { getNewRefreshToken } from '../utils/getNewRefreshToken';
 
 const Index = () => {
   const navigate = useNavigate();
-  const { data, isError, isSuccess } = getTodoList();
-
-  useEffect(() => {
-    if (!apiStorage.token) {
-      navigate('/login');
-    }
-  }, []);
+  const { data, isError, isSuccess, refetch } = getTodoList();
 
   useEffect(() => {
     if (isError) {
-      sessionStorage.clear();
-      navigate('/login');
+      (async () => {
+        const result = await getNewRefreshToken();
+        if (result === false) {
+          sessionStorage.clear();
+          navigate('/login');
+        }
+      })();
     }
   }, [isError]);
